Remove slot-reserving luaH_set
Reserving a slot results in `nil` being left in the table
if the slot isn't used.

This resulted in __newindex creating `nil`s.

This commit replaces the C-derived luaH_set interface with
a new function luaH_setfrom. It also removes the luaH_delete
function in lieu of performing deletions within luaH_set*