Invert the behaviour of luai_apicheck (and adjust api_check accordingly)
This fixes an issue where if an assert fails, rather than the error message,
the string "false" is thrown instead; the root cause of this is that in C,
the error message was automatically generated from the complete expression
passed to assert (condition && message) even if (condition) was false (and
thus fired the assert). In JS, the message has to be passed separately, and
instead (condition && message) was evaluating to false and then producing
`throw new Error(false)`.

Tested: unit tests